handle exception on PrivacyControl
